# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SenAgri B2B",
        "description": "Documentation officielle de l'API SenAgri B2B",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profil"
                ],
                "summary": "Afficher le profil connecté",
                "description": "Retourne les informations du profil de l'utilisateur authentifié.",
                "operationId": "274fbc07758eb418369dc165a7863185",
                "responses": {
                    "200": {
                        "description": "Profil récupéré avec succès"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profil"
                ],
                "summary": "Modifier le profil",
                "description": "Met à jour les informations du profil de l'utilisateur connecté.",
                "operationId": "74599296b1816be41ad4228975229966",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Mamadou Diallo"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Dakar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/profile/photo": {
            "post": {
                "tags": [
                    "Profil"
                ],
                "summary": "Ajouter ou remplacer la photo de profil",
                "description": "Téléverse une nouvelle photo de profil. Si une photo existe déjà, elle est remplacée.",
                "operationId": "5269dbd7d76c309dec93bb67ffd11b94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "description": "Fichier image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo enregistrée avec succès"
                    },
                    "422": {
                        "description": "Fichier invalide"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profil"
                ],
                "summary": "Supprimer la photo de profil",
                "description": "Supprime la photo de profil de l'utilisateur connecté.",
                "operationId": "1d4c0e08394c6007d5e9e2156b6e19d3",
                "responses": {
                    "200": {
                        "description": "Photo supprimée avec succès"
                    },
                    "404": {
                        "description": "Aucune photo de profil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}